results

+++ database/migrations/2022_07_30_145908_create_students_table.php
@@ -16,7 +16,7 @@ class CreateStudentsTable extends Migration
         Schema::create('students', function (Blueprint $table) {
             $table->bigIncrements('student_id');
             $table->string('name')->nullable();
-            $table->integer('mobile_number')->nullable();
+            $table->bigInteger('mobile_number')->nullable();
             $table->timestamps();
         });
     }
